+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - cr2lanelet: Regulatory element to corresponding way relation for lanes (required by autoware)
 - Option to display the aerial image of the current position
 - By default, if no lbdv creditials are stored, the open source credentials are used
+- Possibility to change the geo reference and translate the scenario to another place
 
 ### Changed
 - Remove second conversion option for Lanelet2 and OpenDRIVE conversion example files